Update html proofer setup, Fix tags on Accounts page
